Add ability to keep roshan timer running even when not visible (Action isn't always running but caches the last visible time for future calculation)
Only calculates how much time has passed on action start. This is done via caching the last visible time and deducing how much time has passed since (on start). This calculation happens in the constructor when the action is visible again, so the action isn't ticking or running in the background as intended. This works when changing between stream deck pages or between stream deck software launches.